+++ lib/src/video_player_initializer/file_initializer.dart
@@ -40,6 +40,7 @@ class FileInitializer implements IVideoPlayerInitializerStrategy {
       callbacks: callbacks,
       type: videoSourceConfiguration.videoSourceType,
       globalKeyPlayer: options.globalKeyInitializer,
+      refreshOnUrlReuse: false,
     );
 
     controller.sharedPlayerNotifier.value = Hero(